Misc - 64-bit addresses and lengths

+++ tt_metal/impl/dispatch/device_command.hpp
@@ -95,8 +95,8 @@ public:
     void add_dispatch_write_linear_h(
         uint8_t num_mcast_dests,
         uint32_t noc_xy_addr,
-        uint32_t addr,
-        uint32_t data_sizeB,
+        uint64_t addr,
+        uint64_t data_sizeB,
         const void* data = nullptr,
         uint32_t write_offset_index = 0);
